set up persistance.

+++ domain/src/main/kotlin/com/example/wordcounter/models/Message.kt
@@ -1,3 +1,3 @@
 package com.example.wordcounter.models
 
-data class Message(val id: Long, val message: String)
+data class Message(val id: Int, val message: String)
